+++ tests/terraform/aws/ec2/index.base.ts
@@ -1,6 +1,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as aws from "@pulumi/aws";
 
+// Originally defined at main.tf:5
 const ubuntu = aws.getAmi({
     filters: [
         {
@@ -15,6 +16,7 @@ const ubuntu = aws.getAmi({
     mostRecent: true,
     owners: ["099720109477"],
 });
+// Originally defined at main.tf:21
 const web = new aws.ec2.Instance("web", {
     ami: ubuntu.id,
     instanceType: "t2.micro",